Wired up combo boxes, started importing and exporting

+++ WatchlistUI/GeneralUI.cs
@@ -18,7 +18,7 @@ namespace WatchlistUI
         List<ItemModel> filteredItems = new List<ItemModel>();
         private string sortingType;
         private bool isSortingAscending;
-        private bool isCompleted = false;
+        private bool isCompleted;
         private string _firstListName;        
         private string _secondListName;        
         private string _thirdListName;
@@ -30,15 +30,21 @@ namespace WatchlistUI
         public GeneralUI()
         {
             InitializeComponent();
-            settings = SqliteDataAccess.LoadSettings();
             GetSettings();
+            WireUpForm();
+        }
+
+        public void WireUpForm()
+        {
             LoadItemList(_currentList);
             SetUpDatagridDesign();
-            WireUpCombobox();
+            WireUpComboboxes();
         }
 
         private void GetSettings()
         {
+            settings = SqliteDataAccess.LoadSettings();
+
             _currentList = settings[6].SettingData;
             _firstListName = settings[0].SettingData;
             _firstListCategory = settings[1].SettingData;
@@ -58,24 +64,42 @@ namespace WatchlistUI
             dataGridView1.Columns[4].Width = 40;
         }
 
-        private void WireUpCombobox()
+        private void WireUpComboboxes()
         {
-            if (_currentList == "FirstList")
-            {
-                cmbCurrentList.Text = _firstListName;
-            }
-            else if(_currentList == "SecondList")
-            {
-                cmbCurrentList.Text = _secondListName;
-            }
-            else if (_currentList == "ThirdList")
-            {
-                cmbCurrentList.Text = _thirdListName;
-            }
+            cmbCurrentList.Items.Clear();
+            cmbCategoryFilter.Items.Clear();
+
+            string[] categories;
 
             cmbCurrentList.Items.Add(_firstListName);
             cmbCurrentList.Items.Add(_secondListName);
             cmbCurrentList.Items.Add(_thirdListName);
+
+            if (_currentList == "FirstList")
+            {
+                cmbCurrentList.Text = _firstListName;
+                categories = _firstListCategory.Split(',');
+            }
+            else if(_currentList == "SecondList")
+            {
+                cmbCurrentList.Text = _secondListName;
+                categories = _secondListCategory.Split(',');
+            }
+            else if (_currentList == "ThirdList")
+            {
+                cmbCurrentList.Text = _thirdListName;
+                categories = _thirdListCategory.Split(',');
+            }
+            else
+            {
+                categories = new string[] { };
+            }
+
+
+            foreach (var category in categories)
+            {
+                cmbCategoryFilter.Items.Add(category);
+            }
         }
 
         public void LoadItemList(string currentList)
@@ -91,8 +115,7 @@ namespace WatchlistUI
 
             if (txtTitleFilter.Text != "")
             {
-                //TODO - filtering is case sensitive
-                items = items.Where(x => x.Title.Contains(txtTitleFilter.Text)).ToList();
+                items = items.Where(x => x.Title.ToLower().Contains(txtTitleFilter.Text)).ToList();
             }
             if (txtScoreFilterMin.Text != "")
             {
@@ -110,7 +133,6 @@ namespace WatchlistUI
             {
                 items = items.Where(x => x.Date <= Convert.ToInt32(txtDateFilterMax.Text)).ToList();
             }
-            //TODO - combobox need wireup
             if (cmbCategoryFilter.Text != "")
             {
                 items = items.Where(x => x.Category == cmbCategoryFilter.Text).ToList();
@@ -203,18 +225,22 @@ namespace WatchlistUI
 
         private void btnOnWatch_Click(object sender, EventArgs e)
         {
-            isCompleted = false;
-            btnTransfer.Text = "Transfer to Completed";
-            _currentList = GetSwitchList(_currentList);
-            LoadItemList(_currentList);
-        }
-
-        private void btnCompleted_Click(object sender, EventArgs e)
-        {
-            isCompleted = true;
-            btnTransfer.Text = "Transfer to On watch";
-            _currentList = GetSwitchList(_currentList);
-            LoadItemList(_currentList);
+            if (isCompleted)
+            {
+                isCompleted = false;
+                btnOnWatch.Text = "Switch to Completed";
+                btnTransfer.Text = "Transfer to Completed";
+                _currentList = GetSwitchList(_currentList);
+                LoadItemList(_currentList);
+            }
+            else
+            {
+                isCompleted = true;
+                btnOnWatch.Text = "Switch to On Watch";
+                btnTransfer.Text = "Transfer to On Watch";
+                _currentList = GetSwitchList(_currentList);
+                LoadItemList(_currentList);
+            }
         }
 
         private void btnSettings_Click(object sender, EventArgs e)
@@ -238,9 +264,12 @@ namespace WatchlistUI
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            ItemModel item = (ItemModel)dataGridView1.SelectedRows[0].DataBoundItem;
-            SqliteDataAccess.DeleteFromList(item, _currentList);
-            LoadItemList(_currentList);
+            if (MessageBox.Show("Do you want to delete this?", "Warning!", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+            {
+                ItemModel item = (ItemModel)dataGridView1.SelectedRows[0].DataBoundItem;
+                SqliteDataAccess.DeleteFromList(item, _currentList);
+                LoadItemList(_currentList);
+            }
         }
 
         private void btnSortTitle_Click(object sender, EventArgs e)
@@ -287,6 +316,39 @@ namespace WatchlistUI
 
         private void btnFilter_Click(object sender, EventArgs e)
         {
+            LoadItemList(_currentList);
+        }
+
+        private void cmbCurrentList_SelectionChangeCommitted(object sender, EventArgs e)
+        {
+
+            if (cmbCurrentList.SelectedItem.ToString() == _firstListName)
+            {
+                _currentList = "FirstList";
+            }
+            else if (cmbCurrentList.SelectedItem.ToString() == _secondListName)
+            {
+                _currentList = "SecondList";
+            }
+            else if (cmbCurrentList.SelectedItem.ToString() == _thirdListName)
+            {
+                _currentList = "ThirdList";
+            }
+
+            WireUpForm();
+
+            //cmbCurrentList.;
+        }
+
+        private void btnClearFilter_Click(object sender, EventArgs e)
+        {
+            txtTitleFilter.Text = "";
+            txtScoreFilterMin.Text = "";
+            txtScoreFilterMax.Text = "";
+            txtDateFilterMin.Text = "";
+            txtDateFilterMax.Text = "";
+            cmbCategoryFilter.Text = "";
+
             LoadItemList(_currentList);
         }
     }
